restructure okta auth code to be in plugin folder
