+++ server/plugins/auth-okta/get-stubbed-credentials.js
